backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
dj-database-url==2.3.0
Django==5.0.6
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.8.0
redis==5.2.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.9.0
python-dotenv==1.0.1
